requirements.txt
alembic==1.3.0
attrs==19.3.0
Click==7.0
defusedxml==0.6.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.4
prison==0.1.2
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
